requirements.txt
faker
pyyaml
openai>=1.0.0
pandas
httpx>=0.28.0
flask>=2.3.0
jinja2>=3.1.0
markupsafe>=2.1.0
werkzeug>=2.3.0
gunicorn>=21.0.0
